write out the crontab only if necessary

+++ boot-scripts/remove_failed.sh
@@ -2,7 +2,7 @@
 
 script="/opt/remove_failed.sh"
 
-cat <<"EOF" > $script
+cat <<"EOF" > "$script"
 #!/bin/bash
 
 region="$(curl -sL 169.254.169.254/latest/meta-data/placement/availability-zone | sed '$s/.$//')"
@@ -20,12 +20,12 @@ failed_consul_nodes="$(/bin/consul members | grep failed | awk '{print $1}')"
 EOF
 
 ## evaluate these variables
-cat <<EOF >> $script
+cat <<EOF >> "$script"
 ## lets give it some time
 sleep "${CONSUL_FAILURE_GRACE_PERIOD_IN_SECONDS:-60}"
 EOF
 
-cat <<"EOF" >> $script
+cat <<"EOF" >> "$script"
 for node in $failed_consul_nodes; do
     ## if they are still failed, kill them
     failed_node_name="$(/bin/consul members | grep $node | grep failed | awk '{print $1}')"
@@ -35,11 +35,13 @@ for node in $failed_consul_nodes; do
 done
 EOF
 
-#write out current crontab
-crontab -l > mycron
-#echo new cron into cron file
-echo "* * * * * /bin/bash $script" >> mycron
+#write out current crontab if necessary
+if ! crontab -l | grep -q "$script"; then
+    crontab -l > mycron
+    #echo new cron into cron file
+    echo "* * * * * /bin/bash $script" >> mycron
 
-#install new cron file
-crontab mycron
-rm mycron
+    #install new cron file
+    crontab mycron
+    rm mycron
+fi
